Coding sample in R
Extracts an unstable but relevant set of vars by excluding a stable set of unwanted vars. We never know which variables we'll be passed, but we do know which variables we don't need in the data set. This function uses logic based on that fact and it works because NOT finding a var in a dataset kicks off a warning, but not FINDING a var in a data set causes a program-stopping error. We want warnings, not errors, in this situation.