Sample.Rmd
This is a mid-project snapshot. It shows a process rather than a product.  It is part training manual for my replacement (if ever needed), part code, part code comments, and the beginnings of what will eventually be a data dictionary. 

Its intended audience will need a lot of hand-holding: hence the long explanations.

I have anonymized the variable names and classifications, so bits won't make any sense and the code won't run.